refactor: extract terminal rendering helpers to their own part
terminal_runtime_shell_launches.dart grew past its max-lines baseline
after later shell-env work. move the rendering and platform-support
helpers (xterm theme, fonts, colors, cursor shape, target platform) into
terminal_runtime_rendering.dart, since they were unrelated to shell
launch resolution and only shared the file.
